fix: address build errors
