+++ src/app/api/users/forgot-password/route.ts
@@ -16,7 +16,7 @@ export async function POST(request: NextRequest) {
       }
     );
 
-    let data: any = null;
+    let data = null;
     const ct = response.headers.get("content-type") || "";
     if (ct.includes("application/json")) {
       data = await response.json().catch(() => null);
