+++ src/app/auth/forgot-password/page.tsx
@@ -1,18 +1,22 @@
 import Image from "next/image";
 import { ForgotPasswordForm } from "@/components/forgot-password-form";
+import Link from "next/link";
 
 export default function LoginPage() {
   return (
     <div className="flex min-h-svh flex-col items-center justify-center gap-6 p-6 md:p-10">
       <div className="flex w-full max-w-sm flex-col gap-6">
-        <a href="/" className="flex items-center gap-2 self-center font-medium">
+        <Link
+          href="/"
+          className="flex items-center gap-2 self-center font-medium"
+        >
           <Image
             src="/logo/logo-long.png"
             alt="GoOut Logo"
             width={100}
             height={50}
           />
-        </a>
+        </Link>
         <ForgotPasswordForm />
       </div>
     </div>
