+++ src/app/cart/error.tsx
@@ -15,7 +15,7 @@ export default function CartError({
     <div className="container mx-auto px-4 py-8">
       <div className="max-w-2xl mx-auto text-center">
         <AlertCircle className="mx-auto h-16 w-16 text-red-400 mb-4" />
-        <h1 className="text-2xl font-bold text-gray-900 mb-2">Something went wrong!</h1>
+        <h1 id={`${error}`} className="text-2xl font-bold text-gray-900 mb-2">Something went wrong!</h1>
         <p className="text-gray-600 mb-6">
           We encountered an error while loading your cart. Please try again.
         </p>
