+++ src/app/checkout/error.tsx
@@ -15,7 +15,7 @@ export default function CheckoutError({
     <div className="container mx-auto px-4 py-8">
       <div className="max-w-2xl mx-auto text-center">
         <AlertCircle className="mx-auto h-16 w-16 text-red-400 mb-4" />
-        <h1 className="text-2xl font-bold text-gray-900 mb-2">Checkout Error</h1>
+        <h1 id={`${error}`} className="text-2xl font-bold text-gray-900 mb-2">Checkout Error</h1>
         <p className="text-gray-600 mb-6">
           We encountered an error during checkout. Please try again or return to your cart.
         </p>
