+++ src/components/forgot-password-form.tsx
@@ -74,7 +74,7 @@ export function ForgotPasswordForm({
         <CardHeader className="text-center">
           <CardTitle className="text-xl">Forgot your password?</CardTitle>
           <CardDescription>
-            Enter your email address and we'll send you a link to reset your
+            Enter your email address and we&apos;ll send you a link to reset your
             password
           </CardDescription>
         </CardHeader>
